+++ plugin/nube/database/postgres/pgmodel/model.go
@@ -53,7 +53,7 @@ type FlowNetworkClone struct {
 	CommonUUID
 	CommonName
 	CommonDescription
-	GlobalUUID string `json:"global_uuid,omitempty" gorm:"unique"`
+	GlobalUUID string `json:"global_uuid,omitempty"`
 	ClientId   string `json:"client_id,omitempty"`
 	ClientName string `json:"client_name,omitempty"`
 	SiteId     string `json:"site_id,omitempty"`
@@ -81,7 +81,7 @@ type Consumer struct {
 	CommonName
 	CommonDescription
 	CommonEnable
-	ProducerUUID       string    `json:"producer_uuid,omitempty" gorm:"uniqueIndex;not null"`
+	ProducerUUID       string    `json:"producer_uuid,omitempty"`
 	ProducerThingName  string    `json:"producer_thing_name,omitempty"`
 	ProducerThingUUID  string    `json:"producer_thing_uuid,omitempty"` // this is the remote point UUID
 	ProducerThingClass string    `json:"producer_thing_class,omitempty"`