+++ src/dbhandler/flow_network_clone.go
@@ -5,6 +5,14 @@ import (
 	"github.com/NubeIO/nubeio-rubix-lib-models-go/pkg/v1/model"
 )
 
+func (h *Handler) GetFlowNetworks(args api.Args) ([]*model.FlowNetwork, error) {
+	q, err := getDb().GetFlowNetworks(args)
+	if err != nil {
+		return nil, err
+	}
+	return q, nil
+}
+
 func (h *Handler) GetFlowNetworkClones(args api.Args) ([]*model.FlowNetworkClone, error) {
 	q, err := getDb().GetFlowNetworkClones(args)
 	if err != nil {